AAAAAAAAAAAAAAAAAAAAAAA

+++ LocadoraVeiculo/LocadoraDeVeiculos.WebApiCore/Config/AutoMapperConfig/DescontoProfile.cs
@@ -13,7 +13,7 @@ namespace LocadoraDeVeiculos.WebApi.Config.AutoMapperConfig
                 .ForMember(dest => dest.Tipo, opt => opt.MapFrom(src => src.Tipo));
 
             CreateMap<Desconto, DescontoDetailsViewModel>()
-                .ForMember(dest => dest.Tipo, opt => opt.MapFrom(src => (int)src.Tipo));
+                .ForMember(dest => dest.Tipo, opt => opt.MapFrom(src => src.Tipo));
 
             CreateMap<DescontoCreateViewModel, Desconto>()
                 .ForMember(dest => dest.Tipo, opt => opt.MapFrom(src => (TipoDesconto)src.Tipo));
